Feedback Button and Form in fully-contained component. Added to top-level and now available on every view
