+++ client/src/App.css
@@ -345,7 +345,7 @@ That is 0.71 x 1 */
   translate: 30px;
   position: fixed;
   font-size: medium;
-  bottom: 10%;
+  bottom: 11%;
   right: 0;
   z-index: -1;
   padding-top: .35rem;
@@ -355,18 +355,19 @@ That is 0.71 x 1 */
 }
 
 .feedback-form {
+  background-color: rgb(255, 255, 255);
+  transition:cubic-bezier(0.755, 0.05, 0.855, 0.06);
+  text-align: center;
   border: 1.5px solid rgb(255, 255, 255);
   flex-direction: column;
   position: fixed;
-  background-color: rgb(255, 255, 255);
-  max-width: 250px;
-  min-width: 100px;
-  width: 50%;
-  top: 35%;
-  left: 50%;
+  bottom: 7%;
+  min-width: 250px;
+  max-width: 350px;
+  height: 382px;
+  right: 0;
   box-shadow:
   inset 0 0 0 4px rgb(101, 100, 100);
-  transform: translate(-50%, -20%);
   border-radius: 1rem;
   padding: .85rem;
   z-index: 5; /* Higher than the z-index of the background */
@@ -387,7 +388,6 @@ That is 0.71 x 1 */
   font-size: 22px;
   color: white;
   margin-top: 10px;
-  text-align: left;
 }
 
 .close-feedback-icon {
@@ -407,6 +407,11 @@ That is 0.71 x 1 */
   padding: 0;
   margin-top: 38px;
   z-index: inherit;
+}
+
+.feedback-results-text {
+  padding: 0;
+  margin-top: 100px;
 }
 
 .feedback-button-container {
@@ -453,4 +458,21 @@ That is 0.71 x 1 */
 .close-feedback:hover {
   background-color: #ff0000;
   cursor: pointer;
+}
+
+.close-feedback-complete {
+  background-color: #000000;
+  color: rgb(255, 255, 255);
+  height: 2.6rem;
+  width: 35%;
+  border: none;
+  border-radius: 0.3rem;
+  font-size: 1rem;
+  max-width: 125px;
+  cursor: pointer;
+  margin-top: 50px;
+}
+
+.close-feedback-complete:hover {
+  background-color: #4c4c4c;
 }
